Move sortToBottom implementation to be internal to Column, leaving the GridSorter API untouched
